// projects/aas-lib/src/lib/types/message-entry.ts
/******************************************************************************
 *
 * Copyright (c) 2019-2024 Fraunhofer IOSB-INA Lemgo,
 * eine rechtlich nicht selbstaendige Einrichtung der Fraunhofer-Gesellschaft
 * zur Foerderung der angewandten Forschung e.V.
 *
 *****************************************************************************/

export interface MessageEntry {
    header?: string;
    text: string;
    delay: number;
    autohide: boolean;
    classname?: string;
}